requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.1
fonttools==4.59.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
narwhals==2.1.2
numpy==2.3.2
packaging==25.0
pillow==11.3.0
plotly==6.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
sscws==2.4.4
urllib3==2.5.0
Werkzeug==3.1.3
